Create Storage class to handle data loading and writing.
The new Storage class contains methods and processes that run on
startup, for the handling of data.
